requirements.txt
django
djangorestframework==3.11.0